requirements.txt
transformers
torch
spacy
sentence-transformers
networkx
pyvis
streamlit
PyMuPDF
nltk
openai
python-dotenv